Added Bullet and Ricochet Prefabs
The bullet has a 2D physics material called "Bouncy" other prefabs can add a force with a SCRIPT of a similar name "BouncySurface".

Most prefabs add a small force to the bullet with "BouncySurface" to keep it going. This is not 100% sure if this is necessary but can add a bit of excitement.

The bullet has a preset direction with the "Bullet" script but can be changed later on.